remove unused tests and contracts
